feat: create dto package and refactor responses
