+++ src/main/java/hng_java_boilerplate/squeeze/dto/ResponseMessageDto.java
@@ -1,4 +1,4 @@
-package hng_java_boilerplate.squeeze.util;
+package hng_java_boilerplate.squeeze.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Getter
 @Setter
 @AllArgsConstructor
-public class ResponseMessage {
+public class ResponseMessageDto {
     private String message;
     private int status_code;
 }
